fix: some fixes to tests
