+++ packages/pn-personafisica-webapp/src/navigation/__test__/RequiredAuth.test.tsx
@@ -1,4 +1,4 @@
-import { render } from '@testing-library/react';
+import { render } from "../../__test__/test-utils";
 import * as redux  from '../../redux/hooks';
 import RequireAuth from "../RequiredAuth";
